Fixed back button issue and problem with login flow,Removed Splash Screen, Added Launch theme,Refactoring of some widgets,Signout method repaired,Flow of Navigation fixed
