--- lib/ui/views/login/splash_screen.dart
@@ -1,34 +1,0 @@
-import 'dart:async';
-import 'package:flutter/material.dart';
-import 'package:scan_shot/ui/views/landing.dart';
-
-class SplashScreen extends StatefulWidget {
-  @override
-  _SplashScreenState createState() => _SplashScreenState();
-}
-
-class _SplashScreenState extends State<SplashScreen> {
-  @override
-  void initState() {
-    super.initState();
-    new Future.delayed(
-        const Duration(seconds: 3),
-        () => Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => Landing()),
-            ));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        body: Container(
-          decoration: BoxDecoration(
-              image: DecorationImage(
-                  image: AssetImage("assets/bg.jpg"), fit: BoxFit.cover)),
-        ),
-      ),
-    );
-  }
-}
